added higher order components to handle auth reroutes
added RequireAuth and RequireNoAuth to handle route gaurding in the appropriate places. RequireAuth prevents access to routes that require a jwt
RequireNoAuth prevents access to public pages involved with account creation/validation
