+++ client/src/components/views/Signin.js
@@ -1,8 +1,9 @@
 import React, {Component} from 'react'
 import PropTypes from 'prop-types'
-import {withRouter} from 'react-router-dom'
 import axios from 'axios'
 
+// also includes withRouter
+import RequireNoAuth from '../hoc/RequireNoAuth'
 import {baseRoute} from '../../utils/api'
 
 class Signin extends Component {
@@ -59,4 +60,4 @@ Signin.propTypes = {
   updateAuth: PropTypes.func.isRequired
 }
 
-export default withRouter(Signin)
+export default RequireNoAuth(Signin)
